Decouple EntityFramework from the StandardBehaviors/StandardDataSource. Also further decouple the pure LINQ parts of StandardDataSource into a standalone type QueryableDataSourceBase so that those parts can be reused regardless of the type T in IDataSource<T>.
