+++ src/IntelliTect.Coalesce/Api/Behaviors/SaveKind.cs
@@ -1,0 +1,8 @@
+﻿namespace IntelliTect.Coalesce
+{
+    public enum SaveKind
+    {
+        Create,
+        Update,
+    }
+}
